requirements.txt
alembic==1.7.6
arabic-reshaper==2.1.3
async-generator==1.10
attrs==21.4.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.3
cryptography==36.0.1
Flask==2.0.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
future==0.18.2
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-metadata==4.10.1
importlib-resources==5.4.0
itsdangerous==2.0.1
jdatetime==3.8.2
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
outcome==1.1.0
Pillow==9.0.1
pycparser==2.21
PyMySQL==1.0.2
pyOpenSSL==22.0.0
python-dotenv==0.19.2
requests==2.27.1
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.31
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
Werkzeug==2.0.2
wsproto==1.0.0
WTForms==3.0.1
zipp==3.7.0
